\name{crimeBQR3-package}
\alias{crimeBQR3-package}
\alias{crimeBQR3}
\docType{package}
\title{
\packageTitle{crimeBQR3}
}
\description{
\packageDescription{crimeBQR3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{crimeBQR3}
\packageIndices{crimeBQR3}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{crimeBQR3}

Maintainer: \packageMaintainer{crimeBQR3}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
